Rename / integrate Magic Bait component into existing RidgesideVillage.dll. Since it was just boilerplate, it's been removed.
