--- Ridgeside Development/MagicBaitForRSV/MagicBaitForRSV/manifest.json
@@ -1,20 +1,0 @@
-﻿{
-  "Name": "Magic Bait for Ridgeside Village",
-  "Author": "Stari",
-  "Version": "1.0.0",
-  "Description": "Allows Magic Bait to function with fish added by Ridgeside Village.",
-  "UniqueID": "Rafseazz.MagicBaitForRSV",
-  "MinimumApiVersion": "3.0.0",
-  "UpdateKeys": [ "Nexus:???" ],
-  "EntryDll": "MagicBaitForRSV.dll",
-  "Dependencies": [
-    {
-      "UniqueID": "Rafseazz.RSVCP",
-      "IsRequired": true
-    },
-    {
-      "UniqueID": "spacechase0.JsonAssets",
-      "IsRequired": true
-    }
-  ]
-}
