+++ Ridgeside Development/Ridgeside Village/RidgesideVillage.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31321.278
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MagicBaitForRSV", "MagicBaitForRSV\MagicBaitForRSV.csproj", "{57169DCD-CBCE-46D2-830A-F787743219F4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RidgesideVillage", "RidgesideVillage.csproj", "{57169DCD-CBCE-46D2-830A-F787743219F4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
