+++ Ridgeside Development/Ridgeside Village/RidgesideVillage/RidgesideVillage.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{57169DCD-CBCE-46D2-830A-F787743219F4}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MagicBaitForRSV</RootNamespace>
-    <AssemblyName>MagicBaitForRSV</AssemblyName>
+    <RootNamespace>RidgesideVillage</RootNamespace>
+    <AssemblyName>RidgesideVillage</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -49,12 +49,12 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="manifest.json" />
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="Pathoschild.Stardew.ModBuildConfig">
       <Version>3.3.0</Version>
     </PackageReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="manifest.json" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
