+++ Ridgeside Development/Ridgeside Village/RidgesideVillage/manifest.json
@@ -1,8 +1,8 @@
-{
+﻿{
     "Name": "Ridgeside Village SMAPI",
     "Author": "Rafseazz",
     "Version": "1.2.8",
-    "Description": "Ridgeside Village SMAPI component",
+    "Description": "Ridgeside Village SMAPI and Harmony Component",
     "UniqueID": "Rafseazz.RidgesideVillage",
     "EntryDll": "RidgesideVillage.dll",
     "MinimumApiVersion": "3.0.0",
